Move handlers to app folder
