+++ app/controllers/webhooks_controller.rb
@@ -1,0 +1,3 @@
+class WebhooksController < ApplicationController
+  def webhook; end
+end
