+++ app/handlers/account_disconnected_notification_handler.rb
@@ -1,0 +1,6 @@
+class AccountDisconnectedNotificationHandler < BaseHandler
+  def execute!
+    bot.api.send_message(chat_id: telegram_profile.id,
+                         text: 'Account was disconnected successfully')
+  end
+end
