+++ app/handlers/default_handler.rb
@@ -1,0 +1,5 @@
+class DefaultHandler < BaseHandler
+  def execute!
+    bot.api.send_message(chat_id: telegram_profile.id, text: 'Unknown action')
+  end
+end
