+++ app/handlers/new_ticket_notification_handler.rb
@@ -1,0 +1,30 @@
+class NewTicketNotificationHandler < BaseHandler
+  attr_accessor :ticket
+
+  def initialize(telegram_profile, ticket)
+    super(telegram_profile)
+    @ticket = ticket
+  end
+
+  def execute!
+    bot.api.send_message(chat_id: telegram_profile.id,
+                         text: notification_text,
+                         reply_markup: notification_kayboard)
+  end
+
+  private
+
+  def notification_text
+    "#{ticket.user.full_name} created new ticket " \
+    "for #{ticket.unit_name}\n" \
+    "Title: #{ticket.name}"
+  end
+
+  def notification_kayboard
+    kb = [
+      Telegram::Bot::Types::InlineKeyboardButton
+        .new(text: 'View on RoomPassport', url: company_ticket_url(ticket))
+    ]
+    Telegram::Bot::Types::InlineKeyboardMarkup.new(inline_keyboard: kb)
+  end
+end
