+++ app/jobs/notify_telegram_disconnect_job.rb
@@ -4,6 +4,6 @@ class NotifyTelegramDisconnectJob < ApplicationJob
   def perform(id)
     telegram_profile = TelegramProfile.find(id).decorate
 
-    Handlers::AccountDisconnectedNotificationHandler.new(telegram_profile).execute!
+    AccountDisconnectedNotificationHandler.new(telegram_profile).execute!
   end
 end
