--- telegram/handlers/default_handler.rb
@@ -1,7 +1,0 @@
-module Handlers
-  class DefaultHandler < BaseHandler
-    def execute!
-      bot.api.send_message(chat_id: telegram_profile.id, text: 'Unknown action')
-    end
-  end
-end
